[package_info_plus_windows] move FFI behind a conditional import
Same as for path_provider_windows:
https://github.com/flutter/plugins/commit/25957033132ac17094e82bec58e429fac073a483

Moves the real implementation of package_info_plus_windows behind a
conditional export, instead exporting a stub on platforms that don't
support dart:ffi. This avoids build breakage in web projects that have
transitive dependencies on package_info_plus (and thus
package_info_plus__windows due to manual endorsement).

This will no longer be necessary once
https://github.com/flutter/flutter/issues/52267 is fixed, since only
Windows builds will ever need to have code-level dependency on
package_info_plus_windows.
